Pulled DrawerListAdapter and NavItem out of MainActivity
