--- app/src/main/res/layout/fragment_preferences.xml
@@ -1,16 +1,0 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.looper.loop.PreferencesFragment">
-
-    <!-- NOTE
-        THIS IS A PLACEHOLDER FOR THE PREFERENCES FRAGMENT
-        AND IS NOT CURRENTLY USED
-        -->
-    <!-- TODO: Update blank fragment layout -->
-    <TextView android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="hello world" />
-
-</FrameLayout>
